% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-binary.R
\name{build_linux_binary}
\alias{build_linux_binary}
\title{Build a binary from a local package tree}
\usage{
build_linux_binary(path = ".", image = "rhub/ubuntu-gcc-release",
  platform = NULL, docker_user = "docker")
}
\description{
Build a binary from a local package tree
}

