% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{install_deps}
\alias{install_deps}
\title{Install dependencies, create a new Docker image}
\usage{
install_deps(path, image_id)
}
\arguments{
\item{path}{Path to a tarball.}

\item{image_id}{Docker image id.}
}
\description{
We share the tarball, and use the new remotes::install_deps()
function.
}
\keyword{internal}

