% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{install_system_requirements}
\alias{install_system_requirements}
\title{Create another image, with sysreqs installed}
\usage{
install_system_requirements(image_id, sysreqs)
}
\value{
Id of the new image, or the old image if no sysreqs
  are needed.
}
\description{
Create another image, with sysreqs installed
}
\keyword{internal}
