% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ localbuilder }
\alias{ localbuilder }
\alias{ localbuilder }
\title{ Build R Binary Packages for Multiple Platforms }
\description{
Currently it supports Linux via Docker. It uses R-hub's Docker images to build binary R packages for
  various platforms.
}
