% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{rebuild_cran}
\alias{rebuild_cran}
\title{Build binary packages for all or part of CRAN}
\usage{
rebuild_cran(image = "rhub/ubuntu-gcc-release", docker_user = "docker",
  repo = ".", packages = NULL, compiled_only = TRUE, cleanup = TRUE,
  logfile = new_cran_logfile())
}
\arguments{
\item{image}{The image to use.}

\item{docker_user}{The user to use in the Docker container.}

\item{repo}{The CRAN-like repository that contains the binary packages.}

\item{packages}{The CRAN packages to build. If it is `NULL`, then
all out of date packages that require compilation, will be built.
See also the `compiled_only` argument to force building packages that
do not require compilation.}

\item{compiled_only}{Whether to build only CRAN packages that require
compilation.}

\item{cleanup}{Whether to clean up the Docker containers and images
after the builds.}
}
\description{
Build binary packages for all or part of CRAN
}
