% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-binary.R
\name{build_linux_binary}
\alias{build_linux_binary}
\title{Build a binary from a source package}
\usage{
build_linux_binary(package, image = "rhub/ubuntu-gcc-release",
  docker_user = "docker", repo = ".")
}
\description{
Build a binary from a source package
}

